Add toolchain file for STM32F411
The reason for this toolchain file and main difference from the one
from Rodos is that the options "-nostdlib" and "-nodefaultlibs" are
removed. This was necessary to get rid of the "memcpy not found" kind
of linker errors.

Another important difference is that the HSE value is not specified in
the toolchain file so it can really be used for any STM32F411.
"HSE_VALE" must be set on the command line or in the preset.